txn-wal: Fix subscription leaks in TxnsRead (#27311)
`TxnsRead` maintains a `Vec` of subscriptions over data shards. Any time
an update is made to the txn shard, `TxnsRead` updates all subscriptions
via a channel. If the receiver of the channel is closed, the
subscription is cleaned up.

Previously, we cleaned up the subscriptions via the
`drain_filter_swapping` method. This method creates an iterator which
only removes the element lazily as the iterator is processed.
Unfortunately, the iterator was immediately dropped and never actually
cleaned up subscriptions.

This commit fixes the issue by using `retain` to clean up subscriptions,
which is not an iterator and removes elements immediately.

Fixes #27305