clusterd: add metric for receipt of last envd command
This will make it possible to write an alert to detect clusters that
haven't heard from envd in a while, like we saw in
MaterializeInc/incidents-and-escalations#28. Envd should be continually
chatting with all clusters via the compute protocol, so any stall in
communication is usually indicative of a severe problem.