adapter: correctly report custom types in mz_columns
For columns with a type of list, map, or record, the mz_columns table
was incorrectly reporting the generic psuedotype for the column, even
when the type was a named type in the catalog.

For example, given the following objects:

    CREATE TYPE my_rec AS (x int, y int);
    CREATE TABLE my_tab (f1 my_rec);

the mz_columns table would incorrectly report that column f1 had type
record (OID 2249), rather than type my_rec with a dynamic OID determined
at envd boot.

This commit is a surgical fix that fixes the immediate issue.

In the future, we may want to consider a deeper fix that makes the
conversion between a `ScalarType` and a `pg_repr::Type` less prone to
this sort of bug. That is a much larger refactor, however.

We'll also want to deprecate `mz_columns.type` in the future, for the
reasons noted in the comment in the patch. That too is left to future
work for expediency.