storage,persist: allow callers of persist_source to handle "cannot serve..." errors
Partially fixes #19097

Most callers will panic, as before, but they now include their purpose
in the panic message.

For storage sinks, we now trigger a graceful suspend-and-restart, which
will get rid of at least one scenario where we previously were seeing
this panic in CI.

The machanism of the panic for sinks was:

1. We create the source/sink, sending render commands to the cluster.
2. Cluster does not start rendering yet, for reasons… delays and
   whatnot.
3. We drop the database which immediately drops the source/sinks,
   releases critical since handles and the like.
4. **Now the cluster wakes up and tries to render the sink. This fails
   with a panic!**

With this fix, we would trigger a suspend-and-restart but eventually
notice that the sink has been dropped and not try to re-render its
dataflow.