Update correction peak metrics synchonrously
This commit removes the asynchronous way of updating metrics for the
persist_sink's peak correction buffer sizes between worker steps.
Instead the peak sizes are now updated together with the size delta
metrics from which they are calculated.

This simplifies the correction buffer metrics logic in exchange for
having to perform more work during persist_sink processing. The hope is
that this additional work will not be noticable.