docs: Fix grammar of mat view WITH options
Fixes two issues:
- It looked as if `WITH ()` would be possible.
- It looked as if `ASSERT NOT NULL` and `RETAIN HISTORY` can come
  only in this order, but actually `ASSERT NOT NULL` can be specified
  also after `RETAIN HISTORY` (even both before and after).

Note that the new diagram is also not perfect, because now it looks as
if `RETAIN HISTORY` could be specified more than once. I think we have
to accept this, because it seems impossible to exactly specify using
BNF that these options can be specified in any order, and `ASSERT NOT
NULL` can be specified any number of times, but `RETAIN HISTORY` can
only be specified once.

Also, the new layout seems cleaner overall, especially when I add
REFRESH soon.

Note that I did not add the optional `=` for `ASSERT NOT NULL`. Started
a discussion on this here:
https://materializeinc.slack.com/archives/C063H5S7NKE/p1716888019218749

Plus the commit removes a stale PrPr notice fragment for
ASSERT NOT NULL.