catalog: Listen for updates in transactions
This commit updates the catalog transaction logic such that updates to
the in-memory catalog are made using the `apply_update` function. In
theory this puts us in a state where all updates to the in-memory
catalog that can be derived from the durable catalog, are done via the
`apply_update` function.

Works towards resolving #24844