adapter: let mz_system modify read-only clusters/schemas
This commit gives the `mz_system` user permission to create and drop
items in read-only clusters and schemas. The intention is to have a way
of easily testing the impact of new `mz_introspection` indexes on
production workloads.

`mz_system` is still not allowed to create dataflows reading from
unstable dependencies by default, though this permission can be given
through the `enable_unstable_dependencies` config.