compute: in persist_sink, use WriteHandle to track upper
As the in-code comment describes, using the upper of a persist_source or
a feedback edge is not reliable in a world with multiple instances of a
persist_sink. Especially when one of them is in read-only mode and not
attempting appends, which is when normally the sink would learn of a new
upper and propagate that through the feedback edge.