storage: gracefully restart upsert source on as_of problems
Similar to the fix for sinks in
https://github.com/MaterializeInc/materialize/pull/27323, we now
gracefully restart upsert sources.

Upsert sources suffer from a similar problem to sinks where a drop
command does not make it to the cluster in time while read/since handles
are already dropped by the controller. In those cases, we might now try
and restart a couple of times until the information makes it to the
cluster.

Fixes #27345