storage: differentiate between append-only and incremental storage-managed collections
In support of 0dt upgrades:
https://github.com/MaterializeInc/materialize/issues/27406

This makes previously existing implicit differences in how we handle
storage-managed collections explicit. So that we can be smarter about
how we handle them during read-only mode / 0dt upgrades.

We also fix potential problems around concurrency, by making all moments
where we try and append updates/retractions based on the current shard
state handle upper mismatches.

This is a **pessimization** for memory usage per differential collection
(see below). Differential collections mirror state that we keep in
memory in some form. With my change, we keep an additional copy of that
in the form of the `desired` collection. I do this so that we don’t have
to continually update/consolidate a desired collection by reading from
persist: we only update our view of the world when there is an upper
mismatch, which is rare today. Today, I don’t think this is a problem,
the in-memory state was negligible before, and doubling it is still
negligible. Maybe a hot take!

I think the types of these collections are already locked in, just by
how we currently use them and drive them around. This change it making
those differences _explicit_:

- Append-only: Only accepts blind writes, writes that can be applied at
  any timestamp and don’t depend on current collection contents.
- Pseudo append-only: We treat them largely as append-only collections
  but periodically (currently on bootstrap) retract old updates from
  them. This change fixes a problem where those retractions might not
  line up with what is there, because we didn’t morally do a
  `compare_and_append`.
- Differential: at any given time `t` , collection contents mirrors some
  (small cardinality) state. The cardinality of the collection stays
  constant if the thing that is mirrored doesn’t change in cardinality.
  At steady state, updates always come in pairs of
  retractions/additions.