[persist/sources] Choose a resumption frontier in the source based on the output upper, not the input since (#27314)
- Choose the "as-of" for the source as a function of the output uppers,
not the input since. (While the since is normally in advance of the
uppers, this does not appear to be the case when the source is being
created or shutting down.)
- This is morally similar to how the controller places advances its read
holds on upstream collections based on the progress of the upper in
downstream collections.
- When we open a read handle for a shard, start from whatever since is
available. (And take care to ensure that we don't expose any
insufficiently-advanced data to downstreams.)
 
### Motivation

For: #26940.

The previous code depends on brittle details of the way that the since
is chosen, including leaking a handle briefly to temporarily prevent
compaction. The result is... still brittle, frankly; but it's hopefully
a step towards a less brittle future where the since of Persist shards
is easier to reason about.

### Tips for reviewer

Nightly here: https://buildkite.com/materialize/nightly/builds/7922

This may still be a little janky, but I'd figured I'd better get it up
and let the source experts tell me exactly which parts are the janky
ones.

### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
- <!-- Add release notes here or explicitly state that there are no
user-facing behavior changes. -->