catalog: Simplify bootstrap update sorting
Previously, the state update sorting code was responsible for two
things:

  1. Sorting state updates by dependency.
  2. Grouping/separating state updates that will be batched when
  applied.

This was a bit confusing and broke the principle of single
responsibility. This commit simplifies the code by making the sorting
function only sort the updates. Group and batching is now the
responsibility of the function that applies the updates.