catalog: Always store unfinalized shards (#27436)
Previously, the enable_storage_shard_finalization flag controlled
whether we stored newly deleted but unfinalized shards in the in-memory
catalog. When the flag was disabled, we would not store them in memory.
This could cause the on-disk catalog state to diverge from the in-memory
catalog state which is confusing. Additionally, there are some parts of
the code that assume that if a shard is missing from the in-memory
catalog, it must have been finalized. There is a careful but fragile
dance in the code that prevents leaking shards when
enable_storage_shard_finalization is off. Furthermore, on startup all
unfinalized shards are loaded into the in-memory catalog regardless of
the flag.

This commit simplifies the code by always storing unfinalized shards in
the in-memory catalog.

Maybe but probably not fixes #27196