catalog: Remove public key from catalog state
Previously, SSH connection public keys were stored in memory in the
`CatalogState` struct. However, this information is not stored durably
in the durable catalog. The actual source of truth is an opaque
`SecretsController`. This presents a challenge for having a
multi-subscriber catalog that listens to durable updates. The catalog
has no way of learning about changes to the public keys. The in-memory
public keys will likely become inconsistent after any changes and
present incorrect information to anyone subscribing to the catalog.

This commit fixes the above issue by removing the public keys from
the `CatalogState` struct.

An alternative, and possibly better, approach would be to store the
public keys in the durable catalog directly. Unfortunately, the
public keys are owned by connections, which are a type of item. Items
must be serialized as a CREATE SQL string, so we'd have to find a way
to encode the public key in a `CREATE CONNECTION` string.

Works towards resolving #24844