storage: ensure command Sequencer is borrowed briefly during rendering
The command sequencer is behind an `Rc<RefCell<_>>` and is shared by
both the worker and the dataflows. In #27445 we discovered that there
are cases where the operators need to use the sequencer during their
first schedule, which happens synchronously as part of rendering.

This means that we must make sure to not keep a `borrow_mut` active in
an earlier stack frame so that operators can take the borrow.

Fixes #27445

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>