timely-util: reclaim memory used for stash vec
The current version of the reclock operator maintains a stash vector
whose capacity only ratchets upwards as it's being used. This means that
if the operator processes a spike of input the memory of that vector
will spike too and will never be reclaimed.

This PR recycles the stash buffer every time the operator gets
scheduled. A more complicated buffer management strategy might need to
be considered in the future but for now all benchmarks seems happy with
this change.

Fixes #27425

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>