RTR is not guaranteed in transactions
Real time recency does not play well with interactive transactions where
the system must commit to a timestamp to run statements at upfront but
the set of sources on which the user intends to run queries on is
revealed as the queries come in.

If a source that was not known at the time of the first query of a
transaction is used in a subsequent query we have two options:
1. Run RTR and continue with the transaction if the timestamp we have
   already committed to is compatible (i.e contains all upstream changes)
2. Ignore RTR completely and run all queries at the timestamp determined
   in the beginning of the transaction.

Materialize currently does option 2, which may be surprising to users
that have RTR enabled. We should think about whether option 1 would be
preferable from a product pespective. That said, if the user is using
frequently changing sources then option 1 would almost certainly raise a
timestamp error.

In any case, this PR removes the test of RTR since it is incompatible
with both options at the moment.

Fixes #27393

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>