compute: set length to zero before running user code
When clearing an lgalloc `Array` there is the possibility that the
`Drop` code of the contained type panics. When that happens the stack
will start to unwind, dropping all active values. This process will
eventually reach the `Array` itself, which as part of its `Drop`
implementation will attempt to re-clear itself, which will now rerun the
`Drop` code on potentially uninitialized memory.

This PR fixes this problem by following the same strategy followed by
`std`, which is to set the length of the vector to zero before
attempting any operation that will run potentially panicking user code.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>