catalog: Remove replica statuses from state (#27473)
Previously, cluster replica statuses were stored in memory in the
`CatalogState` struct. However, this information is not stored durably
in the durable catalog, it's ephemeral information. This presents a
challenge for having a multi-subscriber catalog that listens to durable
updates. The catalog has no way of learning about changes to the
replica statuses. The in-memory statuses will likely become
inconsistent after any changes and present incorrect information to
anyone subscribing to the catalog.

This commit fixes the above issue by removing the statuses from
the `CatalogState` struct and moving to to the `Coordinator`. These
statuses are only used in two places:

  1. In the `Coordinator` when processing cluster status update
  messages.
  2. To update the builtin table `mz_cluster_replica_statuses`.

The first usage makes the `Coordinator` a good place to put the
statuses. The second usage should be removed and
`mz_cluster_replica_statuses` should be a `BuiltinSource` instead of
a `BuiltinTable`. That work is saved for a future commit.

Works towards resolving #24844