compute: fix "controller ready" spam
Previously, the `ComputeController::ready` method would return (i.e.
signal readiness) when any of the replicas had failed, the assumption
being that the next call to `process` takes care of rehydrating the
replica. However, that is not how it works anymore: Replica rehydration
is handled as part of the `maintain` work, which only gets invoked when
the maintenance ticker has expired (once per second).

This could lock the coordinator in a ready-process loop for up to one
second at a time when a replica failed. Upon seeing the compute
controller reporting "ready", the coordinator would call its `process`
method, which wouldn't rehydrate the failed replica, so the controller
would remain "ready", causing the coordinator to call `process` over and
over again, until the maintenance ticker expired. The coordinator
handles controller readiness with priority, so it wouldn't get any
useful work done during that time.

This is fixed now by not returning from `Instance::recv` when a replica
connection fails. Returning in this case is not necessary since the
replica gets marked as failed by `recv`, which lets the next `maintain`
invocation pick it up for rehydration.