dbt-materialize: release v1.8.1 (#27523)
<!--
Describe the contents of the PR briefly but completely.

If you write detailed commit messages, it is acceptable to copy/paste
them
here, or write "see commit messages for details." If there is only one
commit
in the PR, GitHub will have already added its commit message above.
-->

### Motivation

Release `dbt-materialize` `v1.8.1`. See `CHANGELOG.md` for an overview
of the release.