compute: ignore peek times in history reduce
This commit makes `ComputeCommandHistory::reduce` ignore the times of
pending peeks when determining the new dataflow as-of frontiers. The
history doesn't attempt to ensure valid read frontiers for dataflow
dependencies, so it's inconsistent that it tries to do so for peek
dependencies. In both cases the compute controller is responsible for
ensuring dependencies are readable at the requested times.

It would also be very surprising (i.e. a bug) for `AllowCompaction`
commands to allow compaction to a frontier _less than_ the previous
dataflow as-of, so this commit also adds a soft assert to catch that
case.