compute: remove the `PeekTarget::id` method
The `PeekTarget::id` method is somewhat dangerous because it doesn't
give an indication of whether the returned ID references a compute
object or a storage object. In the case of MVs, where there is both a
compute and a storage object with the same ID, this can lead to
confusion. Removing the `PeekTarget::id` method and forcing everyone to
destructure the enum to get at the ID makes it less likely for the user
of the ID to assume the wrong collection type.