compute: reconcile logging collections
Upon reconnecting to a replica, the compute controller expects the
logging collections to be uncompacted and readable from the minimum
time. To ensure that we introduce the ability to "pad" traces in the
`TraceManager` by filling up their compacted times with empty contents.
This is implemented as a thin `TraceReader` wrapper that intercepts
accesses to the trace's logical compaction frontier and otherwise passes
requests through to the inner trace.