frontegg-auth: support service app passwords
This commit teaches Materialize to authenticate service app passwords,
as described in #20894.

Under the hood, a service app password is a tenant API token with the
`user` claim set in its metadata. Authenticating with such an API token
functions as a login to the specified user. To prevent impersonating
Frontegg users, the `user` claim must not contain an `@` character.

Database privileges for service app passwords are managed by issuing the
normal `GRANT` and `REVOKE` commands against the user associated with
the app password.

Touches #20894.