compute: remove `Active*` controller types
This commit removes the `ActiveComputeController` and `ActiveInstance`
types which existed solely to wrap compute controller functionality that
needs access to the storage controller.

Now that the shareable `StorageCollections` exists, the
`ComputeController` and `Instance` types can each own an
`Arc<StorageCollections>`, so the `activate` dance becomes unnecessary.

The only place where access to a `&mut StorageController` is still
needed for now is when the `ComputeController` wants to record
introspection updates. This is solved by passing a `&mut
StorageController` into the `ComputeController::process` call and down
to `record_introspection_updates`.