catalog: Make DurableType more ergonomic
Previously, the `DurableType` trait had two generic types, one for the
key and one for the value. This often caused unconstrained type issues
when using this trait that were annoying to work around. This commit
fixes the issue by updating the trait to have associated types instead
of generic types. This also matches the mental model of durable types
nicely; for each types there's a single associated key and value.