catalog: Correctly handle matching updates
The in-memory `CatalogState` maintains denormalized data for certain
object types. For example, items maintain a list of all dependent items
(other items that depend on it). Updates to an object, for example,
updating the privileges on an object, is represented as a retraction of
the object in its previous state, followed by an addition of the object
in its new state, within a single timestamp.

Naively, we would apply these by removing the object from
`CatalogState` when applying the retraction and then re-inserting it
when applying the addition. The issue with this approach is that we
lose all the denormalized state, which can be prohibitively expensive
to regenerate.

This commit resolves this issue via the following process:

  1. When applying retractions, stash all removed objects in memory
  somewhere.
  2. When applying additions, see if there's a matching retraction, and
  if so restore its denormalized state.

Works towards resolving #24844