adapter: pass read_only as an arg when constructing Controller
Before, we were always initializing it to `false` and then allowing
writes via `allow_writes()`. Now we just pass it to the constructor and
avoid the extra mental overhead of that call and where it is called.

The original motivation is that we want to add support for read-only
mode to the storage controller, and there we want to make sure that we
know whether we're in read-only mode as early as possible. For example
because we initialize a bunch of things during `create_collections`
calls, during bootstrap.

With this, we have to make sure that we run any logic that we previously
ran in `allow_writes()` also in `new()`, when `read_only = false`.