txn-wal: Fix as_of_until test downgrade assertion
Previously, the `as_of_until` test in txn-wal was asserting a specific
downgrade behavior when writing/reading to a data shard. Shards do not
guarantee a specific downgrade behavior and may change that behavior
for performance reasons. If the downgrade behavior ever changes, then
that test will start failing. This commit fixes the issue by not
asserting on specific downgrade behavior and only asserting on actual
data messages.