persist: all* columnar encodings, with new `Schema2` trait (#27084)
This PR introduces new `ColumnEncoder`, `ColumnDecoder`, and `Schema2`
traits in Persist, which are very similar to the existing `PartEncoder`,
`PartDecoder`, and `Schema` traits, but the new version does not use the
`Data` trait or `DynStruct` at all. Instead of generics the new traits
rely on enums and `arrow-rs` types directly. I opted to make this
refactor because the more time I've spent in this part of the code base,
I've started to believe the additional complexity the abstraction
introduces is not worth the benefit.

Using the new traits, this PR also introduces a structured columnar
encoder that handles all types of `Datum`s. Importantly (for
performance) the new encoders downcast only once per Part, just the
existing ones. Micro-benchmarks show the encoding and decoding
performance is exactly the same between the two.

One benchmark I added was encoding and decoding JSON. The new structured
encoders are about 2.5x faster than the existing decoding via Protobuf.

#### TODOs

There are still a few very small TODOs in this PR, but I figured it
would be useful to get it up for an initial review since I'm going to
spend some cycles on other work. Those TODOs are:

* Final `Numeric` encoding. I have a PR up
https://github.com/MaterializeInc/materialize/pull/27474 that introduces
a possible encoding for Numeric, but need to iterate once more on it.
Right now we just encoding using `ProtoNumeric`.
* Final `Range` encoding. Haven't put much thought in here, but right
now we're just using `ProtoRange`. `Range`s are not common so it's
probably fine to stick with this, but wanted to call it out.
* More benchmarks.
* Statistics. Still need to implement statistics for all of the new
types.

### Motivation

Progress towards:
https://github.com/MaterializeInc/materialize/issues/24830

Simplify our columnar/stats code. Caveat, it's easier to write code than
it is to read code! I find this refactoring a bit simpler but other
folks might not, which is totally fine!

### Tips for reviewer

This PR is split into a few commits, you should spend the most energy on
1 and 2:

1. New traits, and persist specific implementations for them.
2. New Columnar Datum encoders and decoders.
3. Small supporting bits for the encoders
4. Encoder and decoder for `SourceData`, largely just a wrapper around
the types in 2
5. Test changes, additional benchmarks, and generated files

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - N/a