storage: allow support to configure the wal_sender_timeout
Allow support to configure the wal_sender_timeout for all PostgreSQL
sources in an environment. This would have come in handy on several
occasions to help users override bad upstream settings for the
wal_sender_timeout.

Because this is generally a setting that the DBA should configure, not
the application, we default to not overridding it.

Co-authored-by: Sean Loiselle <sean@materialize.com>