storage: allow support to set TCP configuration on server side too
The existing TCP keepalives only allow the client to notice when the
connection has been silently dropped. If the server does not configure
its own TCP keepalives, it might not notice if the connection is
silently dropped. This is particularly problematic for replication
connections, which hold a lock on replication slots until the server
notices that the connection has been aborted.

Introduce a new `pg_source_tcp_configure_server` parameter that
indicates that the TCP keepalives configuration should be applied to the
server side of the connection as well.

We default the parameter to `false` since the server-side TCP parameters
are generally accepted to be the upstream DBA's responsibility, but
having the ability to set them on a per-environment basis may be useful
when supporting customers.