storage: install read hold for new collection
I discovered a critical bug in the `ALTER SINK` implementation during
the analysis of this nightly test failure [1]. Even though the
controller was creating a read hold for the new collection I simply
forgot to store it in the sink metadata and was instead dropping on the
floor. Unfortunately this wasn't caught during review nor other tests.

The end result is that if a user runs `ALTER SINK` with the current
version of the code they end up with a state where the controller
ensures the frontier of the *old* collection follows the progress of the
sink with the *new* collection. This behavior continuous until
environmentd restart, at which point the controller grabs the correct
read holds.

However, if the new sink starts falling behind nothing will ensure that
the read frontier of the new sinked collection is held back, which may
lead to the sink being permanently stuck.

[1] https://buildkite.com/materialize/nightly/builds/8116#019024cd-e822-48a8-9cb1-3d27b13157f2

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>