Extend MzReplay with container builder
Adds a container builder generic argument to `MzReplay::mz_replay` to allow
callers to express how and what containers they'd like to receive. The
container builder is a parameter on `mz_replay` instead of the trait so
that callers can specify it as part of the function call instead of needing
to specify it as part of the trait. This limits flexibility, but is
acceptable because we're in control over all implementations.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>