testdrive: Ignore unfinalized shards in catalog
Testdrive has a feature where it compares the state of the in-memory
catalog to the state of the on-disk catalog and throws an error if
there is a difference. The set of unfinalized shards in the catalog are
updated asynchronously by background processes. As a result, the value
may legitimately change after fetching the memory catalog but before
fetching the disk catalog, causing the testdrive comparison to fail.
This commit fixes the issue by ignoring the unfinalized shards in the
catalog comparison to ignore false negatives. Unfortunately, we also
end up ignoring true negatives.

Specifically the implementation sets the value of the disk catalog's
unfinalized shards equal to the value of the memory catalog's
unfinalized shards, so that the comparison never fails.

Fixes #27196