Reachability: Use flatcontainer to encode data
This changes the reachability logging implementation to use flatcontainers
to replay the logging stream, and on the input to the pre-arranged data.
It choses not to change the input to the final arrangement to limit the
scope of this PR. This serves as a testebed to see what's needed to get
flatcontainers into more places, but can be merged on its own.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>