bazel: Update BUILD files for `rocksdb` and `librdkafka` (#27699)
This PR updates the build targets for `rocksdb` and `librdkafka` to
align with their respective Rust build scripts, and make sure `cmake` is
linking against the correct libraries.

When getting an end-to-end build of `environmentd` and `clusterd` to
work, we were failing at the last linking step because of missing
symbols. The issue turned out to be that `cmake` was preferring the
version of OpenSSLv3 that existed on my local machine as opposed to the
OpenSSLv1.1.1w that we build with Bazel. We set some specific flags for
`cmake` to prevent it from searching the system, and instead provide an
OpenSSL root path.

For `rocksdb` we were failing to find the symbols from our
`c_api_extension.h` that includes support `WriteBufferManager`. This is
totally an oversight on my part because the `rocksdb` Cmake rules know
nothing about our fork's extensions. In
https://github.com/MaterializeInc/materialize/pull/27698 we upgrade
`rocksdb` to the upstream version that now includes support for the
`WriteBufferManager` which fixes the issue. We also set the same flags
as `librdkafka` to make sure we're linking against the correct external
libraries.

### Motivation

Progress towards
https://github.com/MaterializeInc/materialize/issues/26796

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - N/a