storage: read-only mode for tables
Work towards zero-downtime upgrades: https://github.com/MaterializeInc/materialize/issues/27406

Now, we only register tables in txn-wal when in read-write mode, plus we
register all tables that the storage controller is aware of when coming
out of read-only mode. This latter requires threading a register_ts from
the timestamp oracle through to `allow_writes()`.

As of this change, components further up the stack are not yet aware of
read-only mode, so they will attempt to write and will then error or
panic. Follow-up commits will introduce read-only mode for those
components as well.