Reachability: Use flatcontainer to encode data (#27717)
This changes the reachability logging implementation to use
flatcontainers
to replay the logging stream, and on the input to the pre-arranged data.
It choses not to change the input to the final arrangement to limit the
scope of this PR. This serves as a testebed to see what's needed to get
flatcontainers into more places, but can be merged on its own.

The PR changes behavior in subtle ways. At the moment, the demux
operator in reachability consolidates its outputs, which hides any
sub-tick-interval changes. With flat containers, we cannot yet use the
consolidating container builder, so we might ship more data to the next
operator. The next operator happens to be an arrangement, so it'll
consolidate the data on its input. This makes the situation less bad,
but we might still have some longer-lived allocations.

It does not move data out of lgalloc because only rows in columnation
are backed by lgalloc, but all other types use the default
implementation. The spilled part in reachability is a `Vec<usize>`, so
it currently can't benefit from lgalloc. This means there is no
potential for a regression in this dimension.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [ ] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
- <!-- Add release notes here or explicitly state that there are no
user-facing behavior changes. -->

---------

Signed-off-by: Moritz Hoffmann <mh@materialize.com>